Move static code into helper.go
